:construction: Migrated all user data that will be stored for study into new ViewModel. Added new fragment which will be used to ask users for personal data that are study relevant
