+++ app/src/main/res/layout/fragment_user_data.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".ui.login.LoginFragment">
+
+    <data>
+        <variable
+            name="userDataFragment"
+            type="de.tuchemnitz.armadillogin.ui.welcome.UserDataFragment" />
+        <variable
+            name="armadilloViewModel"
+            type="de.tuchemnitz.armadillogin.model.ArmadilloViewModel" />
+
+        <variable
+            name="studyUserModel"
+            type="de.tuchemnitz.armadillogin.model.StudyUserDataViewModel" />
+    </data>
+
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="16dp">
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </ScrollView>
+</layout>
